Changed naming and code structure to improve readability.
